fix: modularize correctnessText
