+++ client/src/utils/correctness-text.js
@@ -1,0 +1,15 @@
+const correctnessText = (isCorrect, isStreak) => {
+  const { currentStreak } = JSON.parse(localStorage.getItem('birdScore'))
+  if (isStreak && currentStreak !== 2) {
+    return `That's ${currentStreak} in a row! 🦅`
+  }
+  if (isStreak) {
+    return 'New streak! 🐣'
+  }
+  if (isCorrect) {
+    return 'Correct! 👏'
+  }
+  return 'Incorrect. 😬'
+}
+
+export default correctnessText
